requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
googleapis-common-protos==1.51.0
grpcio==1.28.1
grpcio-tools==1.28.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.5
pkg-resources==0.0.0
protobuf==3.11.3
requests==2.23.0
six==1.14.0
snakeviz==2.1.0
tornado==6.0.4
urllib3==1.25.9
Werkzeug==1.0.1
yarl==1.4.2
